LayoutTests/crypto/workers/subtle/aes-postMessage-worker.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../resources/common.js"></script>
</head>
<body>
<script>

description("Test sending aes crypto keys via postMessage to a worker.");

jsTestIsAsync = true;

crypto.subtle.importKey("raw", asciiToUint8Array("16 bytes of key!"), {name: "aes-cbc", length: 128}, true, ["encrypt", "decrypt"]).then(function(localKey) {
    var worker = new Worker("resources/aes-postMessage-worker.js");
    worker.onmessage = function(evt) {
        if (!evt.data.result) {
            testFailed("Check failed in worker: " + evt.data.message);
        } else {
            testPassed("All checks passed in worker");
            key = evt.data.key;
            shouldBe("key.type", "'secret'");
            shouldBe("key.extractable", "true");
            shouldBe("key.algorithm.name", "'AES-CBC'");
            shouldBe("key.algorithm.length", "128");
            shouldBe("key.usages", "['decrypt', 'encrypt']");
        }
        finishJSTest();
    }
    worker.postMessage(localKey);
});
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
