LayoutTests/crypto/workers/subtle/aes-postMessage-worker-expected.txt

Test sending aes crypto keys via postMessage to a worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All checks passed in worker
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ['decrypt', 'encrypt']
PASS bytesToASCIIString(exportedKey) is rawKeyAscii
PASS successfullyParsed is true

TEST COMPLETE
